requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aioredis==2.0.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asteval==1.0.6
astropy==7.0.1
astropy-iers-data==0.2025.5.5.0.38.14
astroquery==0.4.10
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blosc2==3.3.2
caugetch==0.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clipboard==0.0.4
colorama==0.4.6
comm==0.2.2
configparser==7.2.0
contourpy==1.3.1
cryptography==44.0.3
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.6.1
getpass4==0.0.14.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
h11==0.14.0
h5py==3.13.0
habanero==2.2.0
hitran-api==1.2.2.3
hjson==3.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json-tricks==3.17.3
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_git==0.51.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==25.6.0
kiwisolver==1.4.8
llvmlite==0.44.0
lmfit==1.3.3
lxml==5.4.0
magic-filter==1.0.12
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.2
more-itertools==10.7.0
moviepy==2.1.2
mpldatacursor==0.7.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbdime==4.0.2
nbformat==5.10.4
ndindex==1.9.2
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numba==0.61.2
numexpr==2.10.2
numpy==2.2.4
openai==0.28.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
PeakUtils==1.3.5
periodictable==2.0.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.1
proglog==0.1.12
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
publib==0.4.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyerfa==2.0.1.5
Pygments==2.19.1
pyparsing==3.2.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytube==15.0.0
pytz==2025.1
pyvo==1.6.2
PyYAML==6.0.2
pyzmq==26.3.0
radis==0.16
redis==6.0.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tables==3.10.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
torch==2.6.0+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uncertainties==3.2.3
uri-template==1.3.0
urllib3==2.2.0
vulkan==1.3.275.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
yarl==1.20.0
youtube-search==2.1.2
youtube-search-python==1.6.6
yt-dlp==2025.4.30
